Add files via upload
A layer takes input through bottom connections and makes output through top connections. The models are defined in plaintext protocol buffer schema(prototxt). 
The file python_confmat.py must be copied in caffe/examples/mnist to work for the example. Ensure that in the Caffe configuration file, this section is set as follows: WITH_PYTHON_LAYER := 1. This step is essential to ensure that Caffe can recognize Python layers.